move jieba tokenizer and rename corpus_name->embedding_name
