+++ PaddleNLP/paddlenlp/embeddings/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .token_embedding import CorpusName, TokenEmbedding, JiebaEmbeddingTokenizer
+from .token_embedding import CorpusName, TokenEmbedding
